hcCodeChallenge/requirements.txt
anaconda==0.0.1.1
asn1crypto==0.24.0
astroid==1.6.6
autopep8==1.4.4
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
bleach==2.1.3
configparser==3.7.4
cryptography==2.1.4
cycler==0.10.0
decorator==4.3.0
Django==1.11.24
django-cors-headers==3.0.2
djangorestframework==3.9.4
entrypoints==0.2.3
enum34==1.1.6
functools32==3.2.3.post2
future==0.17.1
futures==3.2.0
gyp==0.1
html5lib==1.0.1
idna==2.6
ipaddress==1.0.17
ipykernel==4.8.2
ipython==5.7.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.17
jedi==0.13.3
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
matplotlib==2.2.4
mccabe==0.6.1
mistune==0.8.3
mysqlclient==1.3.10
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.5.0
numpy==1.16.3
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.4
pluggy==0.9.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pycodestyle==2.5.0
pycrypto==2.6.1
pydocstyle==3.0.0
pyflakes==2.1.1
Pygments==2.2.0
pygobject==3.26.1
pylint==1.9.4
pyparsing==2.4.0
python-dateutil==2.8.0
python-jsonrpc-server==0.1.2
python-language-server==0.26.1
pytz==2019.2
pyxdg==0.25
pyzmq==17.0.0
qtconsole==4.3.1
rope==0.14.0
scandir==1.7
scikit-learn==0.20.3
scipy==1.2.1
SecretStorage==2.3.1
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
snowballstemmer==1.2.1
subprocess32==3.5.3
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.2.1
wrapt==1.11.1
wxPython==3.0.2.0
wxPython-common==3.0.2.0
yapf==0.27.0
